+++ src/main/java/com/bm/education/repositories/ModuleTestResultRepository.java
@@ -3,6 +3,6 @@ package com.bm.education.repositories;
 import com.bm.education.models.ModuleTestResult;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ModuleTestResultRepository extends JpaRepository<ModuleTestResult, Integer> {
-    int countByUserId(Integer userId);
+public interface ModuleTestResultRepository extends JpaRepository<ModuleTestResult, Long> {
+    Long countByUserId(Long userId);
 }
