+++ src/main/java/com/bm/education/repositories/OfferRepository.java
@@ -13,19 +13,26 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface OfferRepository extends JpaRepository<Offer, Integer> {
-    List<Offer> findByUserId(Integer userId);
+public interface OfferRepository extends JpaRepository<Offer, Long> {
+    List<Offer> findByUserId(Long userId);
+
     @Query(value = "SELECT o.id, o.topic, o.description, o.created_at, o.user_id," +
             " o.status, o.updated_at, o.response, u.first_name, u.last_name, u.department" +
             " FROM offers o JOIN users u ON o.user_id = u.id WHERE o.status = :status;", nativeQuery = true)
     List<Offer> findByStatus(@Param("status") String status);
-    @NotNull List<Offer> findAll();
-    @NotNull Optional<Offer> findById(@NotNull Integer id);
+
+    @NotNull
+    List<Offer> findAll();
+
+    @NotNull
+    Optional<Offer> findById(@NotNull Long id);
 
     Page<Offer> findByStatus(OfferStatus status, Pageable pageable);
+
     long count();
+
     long countByStatus(OfferStatus status);
 
     @Modifying
-    void deleteByUser_Id(Integer userId);
+    void deleteByUser_Id(Long userId);
 }
