+++ src/main/java/com/bm/education/repositories/UserCourseRepository.java
@@ -8,12 +8,14 @@ import org.springframework.data.jpa.repository.Modifying;
 
 import java.util.List;
 
-public interface UserCourseRepository extends JpaRepository<UserCourses, Integer> {
+public interface UserCourseRepository extends JpaRepository<UserCourses, Long> {
 
     boolean existsByUserAndCourse(User user, Course course);
+
     List<UserCourses> findByUser(User user);
+
     void deleteByUserAndCourse(User user, Course course);
 
     @Modifying
-    void deleteByUser_Id(Integer userId);
+    void deleteByUser_Id(Long userId);
 }
