+++ src/main/java/com/bm/education/repositories/UserProgressRepository.java
@@ -2,7 +2,6 @@ package com.bm.education.repositories;
 
 import com.bm.education.models.UserProgress;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
@@ -10,11 +9,12 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserProgressRepository extends JpaRepository<UserProgress, Integer> {
-    Optional<UserProgress> findByUserIdAndLessonId(int userId, int lessonId);
+public interface UserProgressRepository extends JpaRepository<UserProgress, Long> {
+    Optional<UserProgress> findByUserIdAndLessonId(Long userId, Long lessonId);
+
     @Query(value = "select count(*) from user_progress where user_id = :userId and course_id = :courseId", nativeQuery = true)
-    int totalCompletedLessonByUserId(@Param("userId") int userId, @Param("courseId") int courseId);
+    Long totalCompletedLessonByUserId(@Param("userId") Long userId, @Param("courseId") Long courseId);
 
     @Query("SELECT count(up) FROM UserProgress up WHERE up.user.id = :userId AND up.module.id = :moduleId AND up.completedAt IS NOT NULL")
-    int countByModuleIdAndUserId(@Param("userId") Integer userId, @Param("moduleId") Integer moduleId);
+    Long countByModuleIdAndUserId(@Param("userId") Long userId, @Param("moduleId") Long moduleId);
 }
